requirements.txt.txt
whois
requests
beautifulsoup4
google
validators
Pillow
pytesseract
python-dotenv
